feat: implement MockEventSource and MockEventSourceFactory for testing
